// src/app/models/generales/empleado-dto.ts
export class EmpleadoDTO {
    idEmpleado: number;
    contrato: string;
    nombre: string;
    sueldo: number;
    fechaNacimiento: Date;
    fechaAntiguedad: Date;
    rfc: string;
    curp: string;
    numeroSs: string;
    idDeptoTrabajo: number;
    idEmpresa: number;
    idLugarTrabajo: number;
    idPuestoTrabajo: number;
    idSexo: number;
    numeroEmpleado: string;

    constructor() {
    }
}